So much PID
Wrote PID subsystem for elevator (Elevator.java, DistanceMotorPID.java)
and corresponding commands (MoveElevatorCommand.java) to control the
elevator. The setpoint is actively set, based on the position of the
joystick. max input is 1000, min input is 0. initial setpoint is 0, then
changes to be based on joystick. kP of 0.1, kD = kI = 0. The elevator
guillotines; like literally, fingers could be dismembered.