Added Intake States and Intake Constructor
To clarify: in my "Worked on Intake" update, I meant to say I had given
the Intake Victors for motors (not Solenoids, because those aren't a
motor type). "Worked on Intake" is double-reverted by mistake. In this
update, I added states for the Intake arms and added to the Intake
constructor (using Eyob's elevator test code as a model). I'm not sure
if this is the way states were intended to be used, so if they're being
implemented incorrectly feel free to change/remove them.