Project Documentation
Added LOTS of helpful documentation to the commands controlling the
elevator and the Elevator class itself. Also, the resetPosition method
on the elevator now only sets the conceptual variable to the bottom of
the elevator and does not set the setpoint to the bottom of the
elevator. ResetElevator calls this in it's end method now