Added all requested widgets
Struggling with these terrible docs... (Thanks, FIRST). Elevator
position now literally looks like an elevator screen. It is a series of
progressive "RedGreens" that are lined up vertically. I believe that the
subsystem statuses should be displayed as text and not booleans, as it
is more informative, and so have implemented them this way. I had to
assume that their "path" property is the label that you feed to the
smartdashboard via the robot code (really terrible name for the
property, though). Each of the widgets has a path value which should be
the label passed in (with the exception of the camera).