Made elevator height constants
Elevator height is now stored in a field in Elevator objects.
ElevatorBangBang now takes in those constants as parameters and sets the
Elevator's height field once done. The buttons for elevator movement
have also been changed to reflect the requests of the drivers. Finally,
a lot of unused imports were deleted.